Start db refactor
