--- packages/core/src/db/packs.ts
@@ -1,131 +1,0 @@
-import { Entity, type EntityItem } from 'electrodb';
-import { config, auditAttributes } from './_utils';
-
-export type Pack = EntityItem<typeof packs>;
-export const packs = new Entity(
-	{
-		model: {
-			entity: 'pack',
-			version: '2',
-			service: 'card-app',
-		},
-		attributes: {
-			packId: {
-				type: 'string',
-				required: true,
-			},
-			packTypeId: {
-				type: 'string',
-				required: true,
-			},
-			packTypeName: {
-				type: 'string',
-				required: true,
-			},
-			seasonId: {
-				type: 'string',
-			},
-			username: {
-				type: 'string',
-			},
-			userId: {
-				type: 'string',
-			},
-			cardDetails: {
-				type: 'list',
-				required: true,
-				items: {
-					type: 'map',
-					properties: {
-						instanceId: {
-							type: 'string',
-							required: true,
-						},
-						designId: {
-							type: 'string',
-							required: true,
-						},
-						cardName: {
-							type: 'string',
-							required: true,
-						},
-						cardDescription: {
-							type: 'string',
-							required: true,
-						},
-						imgUrl: {
-							type: 'string',
-							required: true,
-						},
-						rarityId: {
-							type: 'string',
-							required: true,
-						},
-						rarityName: {
-							type: 'string',
-							required: true,
-						},
-						rarityColor: {
-							type: 'string',
-							required: true,
-						},
-						frameUrl: {
-							type: 'string',
-							required: true,
-						},
-						opened: {
-							type: 'boolean',
-							required: true,
-							default: false,
-						},
-						cardNumber: {
-							type: 'number',
-							required: true,
-						},
-						totalOfType: {
-							type: 'number',
-							required: true,
-						},
-					},
-				},
-			},
-			...auditAttributes('pack'),
-		},
-		indexes: {
-			allPacks: {
-				pk: {
-					field: 'pk',
-					composite: [],
-				},
-				sk: {
-					field: 'sk',
-					composite: ['packId'],
-				},
-			},
-			byUsername: {
-				index: 'gsi1',
-				pk: {
-					field: 'gsi1pk',
-					composite: ['username'],
-				},
-				sk: {
-					field: 'gsi1sk',
-					composite: ['packId'],
-				},
-			},
-			byPackId: {
-				index: 'gsi2',
-				collection: 'packsAndCards',
-				pk: {
-					field: 'gsi2pk',
-					composite: ['packId'],
-				},
-				sk: {
-					field: 'gsi2sk',
-					composite: [],
-				},
-			},
-		},
-	},
-	config
-);
