+++ stacks/database.ts
@@ -4,6 +4,43 @@ import { ConfigStack } from './config';
 export function Database({ stack }: StackContext) {
   const config = use(ConfigStack);
 
+  const updatedTable = new Table(stack, 'Database', {
+    fields: {
+      pk: 'string',
+      sk: 'string',
+      gsi1pk: 'string',
+      gsi1sk: 'string',
+      gsi2pk: 'string',
+      gsi2sk: 'string',
+      gsi3pk: 'string',
+      gsi3sk: 'string',
+      gsi4pk: 'string',
+      gsi4sk: 'string',
+    },
+    primaryIndex: {
+      partitionKey: 'pk',
+      sortKey: 'sk',
+    },
+    globalIndexes: {
+      gsi1: {
+        partitionKey: 'gsi1pk',
+        sortKey: 'gsi1sk',
+      },
+      gsi2: {
+        partitionKey: 'gsi2pk',
+        sortKey: 'gsi2sk',
+      },
+      gsi3: {
+        partitionKey: 'gsi3pk',
+        sortKey: 'gsi3sk',
+      },
+      gsi4: {
+        partitionKey: 'gsi4pk',
+        sortKey: 'gsi4sk',
+      },
+    },
+  });
+
   const table = new Table(stack, 'data', {
     fields: {
       pk: 'string',
